filestructure static analysis

+++ contracts/token/ModlVesting.sol
@@ -2,7 +2,7 @@
 pragma solidity ^0.8.7;
 
 import "@openzeppelin/contracts/access/AccessControl.sol";
-import "./IModl.sol";
+import "../interfaces/IModl.sol";
 import "../libraries/Time.sol";
 
 contract ModlVesting is AccessControl {
@@ -11,33 +11,49 @@ contract ModlVesting is AccessControl {
     struct Vesting {
         uint64 start;
         uint64 end;
-        uint amount;
+        uint256 amount;
     }
 
     mapping(address => Vesting) public vesting;
 
-    uint public totalAllocated;
-    uint public ceiling;
+    uint256 public totalAllocated;
+    uint256 public ceiling;
 
     IModl private _modl;
 
-    event Claim(address account, uint amount);
+    event Claim(address account, uint256 amount);
 
-    constructor (IModl modl) {
+    constructor(IModl modl) {
         _modl = modl;
         _grantRole(DEFAULT_ADMIN_ROLE, msg.sender);
-        ceiling = 7500000 * 10 ** 18;
+        ceiling = 7500000 * 10**18;
     }
 
-    function setCeiling(uint newCeiling) external onlyRole(DEFAULT_ADMIN_ROLE) {
-        require(newCeiling >= totalAllocated, "CMERR: Ceiling cannot be less than current allocation.");
+    function setCeiling(uint256 newCeiling)
+        external
+        onlyRole(DEFAULT_ADMIN_ROLE)
+    {
+        require(
+            newCeiling >= totalAllocated,
+            "CMERR: Ceiling cannot be less than current allocation."
+        );
 
         ceiling = newCeiling;
     }
 
-    function update(address account, uint duration, uint amount) external onlyRole(VESTING_MANAGER) {
-        require(account != address(0), "CMERR: account must not be null address.");
-        require(duration > 0 && duration < 315360000, "CMERR: Duration must be greater than 0, and less than 10 years.");
+    function update(
+        address account,
+        uint256 duration,
+        uint256 amount
+    ) external onlyRole(VESTING_MANAGER) {
+        require(
+            account != address(0),
+            "CMERR: account must not be null address."
+        );
+        require(
+            duration > 0 && duration < 315360000,
+            "CMERR: Duration must be greater than 0, and less than 10 years."
+        );
 
         _claim(account);
 
@@ -47,24 +63,33 @@ contract ModlVesting is AccessControl {
         vesting[account].end = uint64(Time.now_u256() + duration);
         vesting[account].amount = amount;
 
-        require(totalAllocated <= ceiling, "CMERR: Cannot allocate more than ceiling.");
+        require(
+            totalAllocated <= ceiling,
+            "CMERR: Cannot allocate more than ceiling."
+        );
     }
 
     function emergencyStop(address account) external onlyRole(VESTING_MANAGER) {
-        require(account != address(0), "CMERR: account must not be null address.");
+        require(
+            account != address(0),
+            "CMERR: account must not be null address."
+        );
 
         totalAllocated = totalAllocated - vesting[account].amount;
         vesting[account].amount = 0;
     }
 
     function claim(address account) external {
-        require(msg.sender == account || hasRole(VESTING_MANAGER, msg.sender), "CMERR: Unauthorized account");
+        require(
+            msg.sender == account || hasRole(VESTING_MANAGER, msg.sender),
+            "CMERR: Unauthorized account"
+        );
 
         _claim(account);
     }
 
     function _claim(address account) internal {
-        uint amount = claimableAmount(account);
+        uint256 amount = claimableAmount(account);
 
         vesting[account].amount -= amount;
         vesting[account].start = Time.now_u64();
@@ -75,7 +100,8 @@ contract ModlVesting is AccessControl {
     }
 
     function claimableAmount(address account) public view returns (uint256) {
-        return Time.windowElapsedValue(
+        return
+            Time.windowElapsedValue(
                 vesting[account].start,
                 vesting[account].end,
                 vesting[account].amount
